+++ src/app/config/static.config.ts
@@ -9,6 +9,10 @@ export const CustomConfig = {
       expiresIn: 0, // wie lange wird das notwendige Cookie gespeichert (in Tagen, oder 0 für die Session)
     },
   },
+  kiConfig: {
+    online: true,
+    mnistPath: 'mnistnumber/model.json',
+  },
   paths: {
     lnkLanding: '/',
     lnkOffers: '/kurse',